add tests for vm-import
