+++ extensions/topsql/src/lib.rs
@@ -7,6 +7,11 @@ mod shutdown;
 mod topology;
 mod upstream;
 
+// Since topsql is highly associated with vm_import,
+// expose the event builder to vm_import for test.
+#[cfg(feature = "vm-test")]
+pub use upstream::parser;
+
 // #[cfg(test)]
 // mod tests {
 //     use std::fs::read;
