+++ extensions/topsql/src/topology/fetch/pd.rs
@@ -106,10 +106,3 @@ impl<'a> PDTopologyFetcher<'a> {
         Ok(members_resp)
     }
 }
-
-#[cfg(test)]
-mod tests {
-
-    #[tokio::test]
-    async fn fetch_health() {}
-}
