+++ extensions/vm-import/Cargo.toml
@@ -19,3 +19,6 @@ async-trait = { version = "0.1.56", default-features = false }
 toml = { version = "0.5.9", default-features = false }
 typetag = { version = "0.2.1", default-features = false }
 hyper = { version = "0.14.20", default-features = false, features = ["client", "runtime", "http1", "http2", "server", "stream"] }
+
+[dev-dependencies]
+topsql = { path = "../topsql", features = ["vm-test"] }
