+++ extensions/vm-import/src/lib.rs
@@ -1,207 +1,9 @@
 #[macro_use]
 extern crate tracing;
 
-use std::io::Write;
-
-use bytes::{BufMut, Bytes, BytesMut};
-use flate2::write::GzEncoder;
-use flate2::Compression;
-use futures_util::{FutureExt, SinkExt};
-use http::{Request, Uri};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use vector::config::{AcknowledgementsConfig, GenerateConfig, Input, SinkConfig, SinkDescription};
-use vector::event::Event;
-use vector::http::HttpClient;
-use vector::sinks::util::http::{BatchedHttpSink, HttpEventEncoder, HttpSink};
-use vector::sinks::util::{
-    BatchConfig, BoxedRawValue, JsonArrayBuffer, SinkBatchSettings, TowerRequestConfig,
-};
-use vector::tls::{TlsConfig, TlsSettings};
-use vector::{config, sinks};
-
-inventory::submit! {
-    SinkDescription::new::<VMImportConfig>("vm_import")
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct VMImportConfig {
-    pub endpoint: String,
-    pub tls: Option<TlsConfig>,
-
-    #[serde(default)]
-    pub request: TowerRequestConfig,
-    #[serde(default)]
-    pub batch: BatchConfig<VMImportDefaultBatchSettings>,
-}
-
-#[derive(Clone, Copy, Debug, Default)]
-pub struct VMImportDefaultBatchSettings;
-
-impl SinkBatchSettings for VMImportDefaultBatchSettings {
-    const MAX_EVENTS: Option<usize> = Some(1_000);
-    const MAX_BYTES: Option<usize> = None;
-    const TIMEOUT_SECS: f64 = 1.0;
-}
-
-impl GenerateConfig for VMImportConfig {
-    fn generate_config() -> toml::Value {
-        let sample_url = "http://127.0.0.1:8428/api/v1/import";
-
-        toml::Value::try_from(Self {
-            tls: Default::default(),
-            batch: Default::default(),
-            request: Default::default(),
-
-            endpoint: sample_url.to_owned(),
-        })
-        .unwrap()
-    }
-}
-
-#[async_trait::async_trait]
-#[typetag::serde(name = "vm_import")]
-impl SinkConfig for VMImportConfig {
-    async fn build(
-        &self,
-        cx: config::SinkContext,
-    ) -> vector::Result<(sinks::VectorSink, sinks::Healthcheck)> {
-        let endpoint = self.endpoint.parse::<Uri>()?;
-        let tls_settings = TlsSettings::from_options(&self.tls)?;
-        let batch_settings = self.batch.into_batch_settings()?;
-        let request_settings = self.request.unwrap_with(&Default::default());
-
-        let client = HttpClient::new(tls_settings, cx.proxy())?;
-        let sink = VMImportSink::new(endpoint.clone());
-        let buffer = JsonArrayBuffer::new(batch_settings.size);
-
-        let sink = BatchedHttpSink::new(
-            sink,
-            buffer,
-            request_settings,
-            batch_settings.timeout,
-            client.clone(),
-        )
-        .sink_map_err(|e| error!(message = "VM import sink error.", %e));
-        let hc = healthcheck(endpoint, client).boxed();
-
-        Ok((sinks::VectorSink::from_event_sink(sink), hc))
-    }
-
-    fn input(&self) -> Input {
-        Input::log()
-    }
-
-    fn sink_type(&self) -> &'static str {
-        "vm_import"
-    }
-
-    fn acknowledgements(&self) -> Option<&AcknowledgementsConfig> {
-        None
-    }
-}
-
-async fn healthcheck(endpoint: Uri, client: HttpClient) -> vector::Result<()> {
-    let request = http::Request::get(endpoint).body(hyper::Body::empty())?;
-    let response = client.send(request).await?;
-    let status = response.status();
-    status
-        .is_success()
-        .then_some(())
-        .ok_or_else(move || sinks::HealthcheckError::UnexpectedStatus { status }.into())
-}
-
-struct VMImportSinkEventEncoder;
-
-impl HttpEventEncoder<Value> for VMImportSinkEventEncoder {
-    fn encode_event(&mut self, event: Event) -> Option<Value> {
-        event.try_into_log().and_then(|mut log| {
-            let labels = log.remove("labels")?.into_object()?;
-            let mut target_metric = serde_json::Map::with_capacity(labels.len());
-            for (key, value) in labels {
-                let target_value =
-                    Value::String(String::from_utf8_lossy(value.as_bytes()?.as_ref()).to_string());
-                target_metric.insert(key, target_value);
-            }
-
-            let src_timestamps = log.remove("timestamps")?;
-            let timestamps = src_timestamps.as_array()?;
-            let mut target_timestamps = Vec::with_capacity(timestamps.len());
-            for timestamp in timestamps {
-                let datetime = timestamp.as_timestamp()?;
-                target_timestamps.push(Value::Number(serde_json::Number::from(
-                    datetime.timestamp_millis(),
-                )));
-            }
-
-            let src_values = log.remove("values")?;
-            let values = src_values.as_array()?;
-            let mut target_values = Vec::with_capacity(values.len());
-            for value in values {
-                let number = value.as_float()?;
-                target_values.push(Value::Number(serde_json::Number::from_f64(
-                    *number.as_ref(),
-                )?));
-            }
-
-            let mut target_map = serde_json::Map::with_capacity(3);
-            target_map.insert(
-                "metric".to_owned(),
-                serde_json::Value::Object(target_metric),
-            );
-            target_map.insert(
-                "timestamps".to_owned(),
-                serde_json::Value::Array(target_timestamps),
-            );
-            target_map.insert("values".to_owned(), serde_json::Value::Array(target_values));
-
-            Some(Value::Object(target_map))
-        })
-    }
-}
-
-#[derive(Clone)]
-struct VMImportSink {
-    endpoint: Uri,
-}
-
-impl VMImportSink {
-    const fn new(endpoint: Uri) -> Self {
-        Self { endpoint }
-    }
-}
-
-#[async_trait::async_trait]
-impl HttpSink for VMImportSink {
-    type Input = Value;
-    type Output = Vec<BoxedRawValue>;
-    type Encoder = VMImportSinkEventEncoder;
-
-    fn build_encoder(&self) -> Self::Encoder {
-        VMImportSinkEventEncoder {}
-    }
-
-    async fn build_request(&self, events: Self::Output) -> vector::Result<Request<Bytes>> {
-        let buffer = BytesMut::new();
-        let mut w = GzEncoder::new(buffer.writer(), Compression::default());
-
-        for event in events {
-            w.write_all(event.get().as_bytes())
-                .expect("Writing to Vec can't fail");
-            w.write_all(b"\n").expect("Writing to Vec can't fail");
-        }
-        let body = w
-            .finish()
-            .expect("Writing to Vec can't fail")
-            .into_inner()
-            .freeze();
-
-        let builder = Request::post(self.endpoint.clone()).header("Content-Encoding", "gzip");
-        let request = builder.body(body).unwrap();
-
-        Ok(request)
-    }
-}
+mod config;
+mod encoder;
+mod sink;
 
 // #[cfg(test)]
 // mod tests {
